+++ internal/infrastructure/controllers/user_update_email_test.go
@@ -40,7 +40,7 @@ func TestUpdateUserEmailController_Handle(t *testing.T) {
 			routeParamID: "1",
 			requestBody:  validBody,
 			setupMock: func(m *mockApp.MockUpdateUserEmailUseCase) {
-				m.EXPECT().Execute(mock.Anything, 1, mock.MatchedBy(func(e *domain.Email) bool {
+				m.EXPECT().Execute(mock.Anything, int64(1), mock.MatchedBy(func(e *domain.Email) bool {
 					return e != nil && e.Value() == "new@example.com"
 				})).Return(nil)
 			},
@@ -53,7 +53,7 @@ func TestUpdateUserEmailController_Handle(t *testing.T) {
 				Email: &emptyEmail,
 			},
 			setupMock: func(m *mockApp.MockUpdateUserEmailUseCase) {
-				m.EXPECT().Execute(mock.Anything, 1, (*domain.Email)(nil)).Return(nil)
+				m.EXPECT().Execute(mock.Anything, int64(1), (*domain.Email)(nil)).Return(nil)
 			},
 			expectedStatus: http.StatusNoContent,
 		},
@@ -96,7 +96,7 @@ func TestUpdateUserEmailController_Handle(t *testing.T) {
 			routeParamID: "1",
 			requestBody:  validBody,
 			setupMock: func(m *mockApp.MockUpdateUserEmailUseCase) {
-				m.EXPECT().Execute(mock.Anything, 1, mock.Anything).Return(domain.ErrUserNotFound)
+				m.EXPECT().Execute(mock.Anything, int64(1), mock.Anything).Return(domain.ErrUserNotFound)
 			},
 			expectedStatus: http.StatusNotFound,
 			expectedBody:   domain.ErrUserNotFound.Error(),
@@ -106,7 +106,7 @@ func TestUpdateUserEmailController_Handle(t *testing.T) {
 			routeParamID: "1",
 			requestBody:  validBody,
 			setupMock: func(m *mockApp.MockUpdateUserEmailUseCase) {
-				m.EXPECT().Execute(mock.Anything, 1, mock.Anything).Return(domain.ErrUserEmailAlreadyExists)
+				m.EXPECT().Execute(mock.Anything, int64(1), mock.Anything).Return(domain.ErrUserEmailAlreadyExists)
 			},
 			expectedStatus: http.StatusConflict,
 			expectedBody:   domain.ErrUserEmailAlreadyExists.Error(),
@@ -116,7 +116,7 @@ func TestUpdateUserEmailController_Handle(t *testing.T) {
 			routeParamID: "1",
 			requestBody:  validBody,
 			setupMock: func(m *mockApp.MockUpdateUserEmailUseCase) {
-				m.EXPECT().Execute(mock.Anything, 1, mock.Anything).Return(errors.New("db failed"))
+				m.EXPECT().Execute(mock.Anything, int64(1), mock.Anything).Return(errors.New("db failed"))
 			},
 			expectedStatus: http.StatusInternalServerError,
 			expectedBody:   "internal server error",
